refactor: reorg ShipVerification.lua to Ship/
